+++ core/bench/src/actors/consumer/client/interface.rs
@@ -23,8 +23,8 @@ use crate::actors::{ApiLabel, BatchMetrics, BenchmarkInit};
 
 #[derive(Debug, Clone)]
 pub struct BenchmarkConsumerConfig {
-    pub consumer_id: u32,
-    pub consumer_group_id: Option<u32>,
+    pub consumer_id: usize,
+    pub consumer_group_id: Option<usize>,
     pub stream_id: String,
     pub messages_per_batch: BenchmarkNumericParameter,
     pub warmup_time: IggyDuration,
