+++ core/binary_protocol/src/cli/binary_segments/delete_segments.rs
@@ -31,7 +31,7 @@ impl DeleteSegmentsCmd {
     pub fn new(
         stream_id: Identifier,
         topic_id: Identifier,
-        partition_id: u32,
+        partition_id: usize,
         segments_count: u32,
     ) -> Self {
         Self {
