+++ core/binary_protocol/src/client/binary_clients/message_client.rs
@@ -31,7 +31,7 @@ pub trait MessageClient {
         &self,
         stream_id: &Identifier,
         topic_id: &Identifier,
-        partition_id: Option<u32>,
+        partition_id: Option<usize>,
         consumer: &Consumer,
         strategy: &PollingStrategy,
         count: u32,
@@ -55,7 +55,7 @@ pub trait MessageClient {
         &self,
         stream_id: &Identifier,
         topic_id: &Identifier,
-        partition_id: u32,
+        partition_id: usize,
         fsync: bool,
     ) -> Result<(), IggyError>;
 }
