+++ core/binary_protocol/src/client/binary_clients/topic_client.rs
@@ -45,7 +45,7 @@ pub trait TopicClient {
         &self,
         stream_id: &Identifier,
         name: &str,
-        partitions_count: u32,
+        partitions_count: usize,
         compression_algorithm: CompressionAlgorithm,
         replication_factor: Option<u8>,
         message_expiry: IggyExpiry,
