+++ core/common/src/commands/topics/mod.rs
@@ -24,4 +24,4 @@ pub mod purge_topic;
 pub mod update_topic;
 
 const MAX_NAME_LENGTH: usize = 255;
-const MAX_PARTITIONS_COUNT: u32 = 1000;
+const MAX_PARTITIONS_COUNT: usize = 1000;
